requirements/app.txt
# pin dependencies with '=='
# make sure they are installable on all systems your plugin should support

# check newest API version: https://pypi.org/project/galaxy.plugin.api/
galaxy.plugin.api==0.62

